+++ code/ui.datagrid/test/SAMPLE/getFunc.ts
